Re-enabled dynamic file name upload and download
Reenabled uploading of files based on file name when chosen and refreshing the table after uploading.

Can also curl to the private bucket and also to add the "encrypted" symmetric key to that particular file. Have handled spaces in the file name. Will add more handling if the need arises